Move `ExternalLinks` into `toc_parser`
toc_source_graph doesn't depend on toc_syntax anymore
