+++ compiler/toc_source_graph/Cargo.toml
@@ -7,7 +7,6 @@ edition = "2021"
 
 [dependencies]
 toc_span = { path = "../toc_span" }
-toc_syntax = { path = "../toc_syntax" }
 
 indexmap = { workspace = true }
 petgraph = { version = "0.6.0", default-features = false, features = ["stable_graph"] }
