+++ compiler/toc_source_graph/src/lib.rs
@@ -1,46 +1,15 @@
 use std::collections::HashMap;
 
-use indexmap::{IndexMap, IndexSet};
+use indexmap::IndexSet;
 use petgraph::{
     graph::NodeIndex,
     stable_graph::StableDiGraph,
     visit::{DfsPostOrder, Visitable},
 };
 use toc_span::FileId;
-use toc_syntax::{ast, AstPtr};
 
 #[cfg(test)]
 mod test;
-
-/// What other files each [`ast::ExternalRef`] refers to
-#[derive(Debug, Default, Clone, PartialEq, Eq)]
-pub struct ExternalLinks {
-    links: IndexMap<AstPtr<ast::ExternalRef>, FileId>,
-}
-
-impl ExternalLinks {
-    /// Binds an external link to a file
-    pub fn bind(&mut self, link: AstPtr<ast::ExternalRef>, to: FileId) {
-        self.links.insert(link, to);
-    }
-
-    /// Looks up which file an external link might refer to
-    pub fn links_to(&self, link: AstPtr<ast::ExternalRef>) -> Option<FileId> {
-        self.links.get(&link).copied()
-    }
-
-    /// Iterator over all of the other files each [`ast::ExternalRef`] in a file refers to
-    pub fn all_links(&self) -> impl Iterator<Item = FileId> + '_ {
-        self.links.values().copied()
-    }
-
-    /// Iterator over all of the other files each importable [`ast::ExternalRef`] in a file refers to
-    pub fn import_links(&self) -> impl Iterator<Item = FileId> + '_ {
-        self.links.iter().flat_map(|(k, v)| {
-            (k.syntax_node_ptr().kind() == toc_syntax::SyntaxKind::ExternalItem).then_some(*v)
-        })
-    }
-}
 
 type LibraryGraph = StableDiGraph<FileId, ()>;
